Reload queries when env changes
