+++ packages/toolpad-core/src/jsServerRuntime.tsx
@@ -15,17 +15,18 @@ function evalExpressionInContext(
   }
 }
 
-export async function createServerJsRuntime(): Promise<JsRuntime> {
+export function createServerJsRuntime(env?: Record<string, string | undefined>): JsRuntime {
   return {
+    getEnv() {
+      if (env) {
+        return env;
+      }
+      throw new Error(`Env variables are not supported in this context`);
+    },
     evaluateExpression: (code, globalScope) => evalExpressionInContext(code, globalScope),
   };
 }
 
 export function useServerJsRuntime(): JsRuntime {
-  return React.useMemo(
-    () => ({
-      evaluateExpression: (code, globalScope) => evalExpressionInContext(code, globalScope),
-    }),
-    [],
-  );
+  return React.useMemo(() => createServerJsRuntime(), []);
 }
