+++ packages/toolpad-core/src/types.ts
@@ -454,6 +454,7 @@ export type Serializable =
   | ((...args: Serializable[]) => Serializable);
 
 export interface JsRuntime {
+  getEnv(): Record<string, string | undefined>;
   evaluateExpression(code: string, globalScope: Record<string, unknown>): BindingEvaluationResult;
 }
